+++ src/handlers/buttons/toggle_role/index.ts
@@ -9,14 +9,14 @@ import type { BaseBot } from '@bot';
 import { ButtonHandler } from '@button';
 
 export default class toggle_role extends ButtonHandler {
-    public override async execute(interaction: ButtonInteraction, _bot: BaseBot): Promise<void> {
+    public override async execute(interaction: ButtonInteraction, bot: BaseBot): Promise<void> {
         const member = interaction.member as GuildMember;
         const guild = interaction.guild as Guild;
         const roleId = interaction.customId.split("|")[1] ?? '';
         const role = guild.roles.cache.get(roleId);
         if (!role) {
             await interaction.reply({
-                content: `Role not found.`,
+                content: bot.translator?.t('replies:toggle_role.role_not_found') ?? '',
                 flags: MessageFlags.Ephemeral
             });
             return;
@@ -26,13 +26,13 @@ export default class toggle_role extends ButtonHandler {
         if (member.roles.cache.has(roleId)) {
             await member.roles.remove(roleId);
             await interaction.reply({
-                content: `Removed role ${role.name}.`,
+                content: bot.translator?.t('replies:toggle_role.removed', { role: role.name }) ?? '',
                 flags: MessageFlags.Ephemeral
             });
         } else {
             await member.roles.add(roleId);
             await interaction.reply({
-                content: `Added role ${role.name}.`,
+                content: bot.translator?.t('replies:toggle_role.added', { role: role.name }) ?? '',
                 flags: MessageFlags.Ephemeral
             });
         }